python/requirements.txt
mysql-connector-python
sqlalchemy
python-dotenv
pymysql